+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Unreleased
+
 ## 3.0.0-beta.1
 
 * Fix: Associate spans and events when it throws (#219)