+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Fix: Correctly add the proguard UUID output directory to the source set (#226)
+* Feature: Make the ignoreXXX properties in SentryPluginExtension sets (#225)
 * Expose SentryPluginExtension.tracingInstrumentation (#229)
 
 ## 3.0.0-beta.1