+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Feature: Make the ignoreXXX properties in SentryPluginExtension sets (#225)
 * Expose SentryPluginExtension.tracingInstrumentation (#229)
 * Ref: Change Room queries description to Dao class name (#232)
-- Fix: Log broken bytecode when build fails (#233)
+* Fix: Log broken bytecode when build fails (#233)
 
 ## 3.0.0-beta.1