+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 3.0.0-beta.2
 
 * Fix: Correctly add the proguard UUID output directory to the source set (#226)
 * Feature: Make the ignoreXXX properties in SentryPluginExtension sets (#225)